# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для CV-приложения: теги, технологии, история работы, образование.",
        "title": "CV Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3333",
    "basePath": "/api",
    "paths": {
        "/hero": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hero"
                ],
                "summary": "Получить hero-секцию",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HeroDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
        "/about-me": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "about-me"
                ],
                "summary": "Получить секцию «О себе»",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AboutMeDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "about-me"
                ],
                "summary": "Обновить секцию «О себе»",
                "parameters": [
                    {
                        "description": "Данные секции",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "bioParagraphs": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "hobbies": {
                                            "type": "string"
                                        },
                                        "note": {
                                            "type": "string"
                                        },
                                        "technologyIds": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AboutMeDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
        "/download-cv": {
            "get": {
                "description": "Генерирует актуальное резюме из данных БД и возвращает PDF-файл",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "cv"
                ],
                "summary": "Скачать CV в PDF",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
...
        "/edu/{eduID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "education"
                ],
                "summary": "Получить запись образования по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи образования",
                        "name": "eduID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EducationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
...
        "/tag/{tagID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Получить тег по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TagDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
...
        "/tech/{techID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technologies"
                ],
                "summary": "Получить технологию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID технологии",
                        "name": "techID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TechnologyWithTagsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        },
...
        "/wh/{whID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work-history"
                ],
                "summary": "Получить запись истории работы по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи истории работы",
                        "name": "whID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WorkHistoryWithTechnologiesDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apierrors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.HeroDTO": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "greeting": {
                    "type": "string"
                },
                "pitch": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.AboutMeDTO": {
            "type": "object",
            "properties": {
                "bioParagraphs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hobbies": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TechnologyWithTagsDTO"
                    }
                }
            }
        },
...
        "dto.EducationDTO": {
            "type": "object",
            "properties": {
                "course": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}